+++ server/QuizBuilder.Domain.Model/Default/Attempts/QuizAttempt.cs
@@ -15,7 +15,7 @@ namespace QuizBuilder.Domain.Model.Default.Attempts {
 
 		public string Data { get; set; }
 
-		public double Result { get; set; }
+		public decimal Result { get; set; }
 
 	}
